Dockerize discord-bot

+++ discord-bot/src/index.ts
@@ -3,8 +3,6 @@ import ready from './listeners/ready';
 import interactionCreate from './listeners/interaction_create';
 import 'dotenv/config';
 import readline from 'readline';
-readline.emitKeypressEvents(process.stdin);
-process.stdin.setRawMode(true);
 
 console.log("Bot is starting...");
 const token = process.env['TERRARIA_BOT_TOKEN'];
@@ -17,9 +15,14 @@ interactionCreate(client);
 
 client.login(token);
 
-process.stdin.on('keypress', (str, key) => {
-  if (key.name === 'q') {
-    client.destroy();
-    process.exit();
-  }
-});
+if (process.stdin.isTTY) {
+  readline.emitKeypressEvents(process.stdin);
+  process.stdin.setRawMode(true);
+
+  process.stdin.on('keypress', (str, key) => {
+    if (key.name === 'q') {
+      client.destroy();
+      process.exit();
+    }
+  });
+}
